Restructure Pong for game-end screen. Minor refactoring. Closes #10
Game states were changed to an enum instead of const integers. In several
places, switch statements on enum values were inserted to replace if/else
blocks.

For the game end screen, the game will wait 5 seconds and then return
to the registration screen (UNTESTED). If a player is lost and someone
registers in their place, the game will immediately return to play.